refactor(client): add hook for loading initial data
